+++ python/options.py
@@ -1098,7 +1098,9 @@ class Options(_options_impl.OptionsDict):
                          anchor_return_period=1,
                          connection_type=enums.ConnectionType.Always.value,
                          sync_pattern=enums.SyncPattern.Full.value,
-                         available_memory_proportion={})
+                         available_memory_proportion={},
+                         meanAccumulationAndReplicationReductionStrategy=enums.
+                         MeanReductionStrategy.Post)
         path = os.environ.get("POPTORCH_CACHE_DIR", "")
         if path:
             logger.info("POPTORCH_CACHE_DIR is set: setting cache path to %s",